Moving JS templates into private module
I'm doing this because I'm going to implement at least one chart type and this
file will keep ballooning in size. Because this package is intended to be installed via
PIP, there is no specific benefit to keeping it to one file.
